Write DartSingleProblemPlanner. Change ConfigurationToConfigurationPlanner.
